Batch Testing + Batch Result Output Complete
Added batched testing as well as result output and formatting as .csv's. Tests are also now generated as .json, while this doesn't make much of a functional difference from .txt, it's more correct.

Results are stored in the directory the batch is taken from, and the algorithm in use for that run is appended to the results name. Re-doing a test with the same algorithm will overwrite the previous file, but this should result in no changes anyway.

+++ SchedSim/src/simulator/PCB.java
@@ -2,38 +2,38 @@ package simulator;
 
 public class PCB {
 	private String name;
-	private float burstTime;
-	private float arrivalTime;
-	private float remainingTime;
-	private float executed;
-	private float beginTime;
-	private float waitTime;
-	private float turnaroundTime;
-	private float responseTime;
-	private float completionTime;
+	private int burstTime;
+	private int arrivalTime;
+	private int remainingTime;
+	private int executed;
+	private int beginTime;
+	private int waitTime;
+	private int turnaroundTime;
+	private int responseTime;
+	private int completionTime;
 	
-	public float getArrivalTime() { return this.arrivalTime; }
-	public float getBeginTime() { return this.beginTime; }
-	public float getBurstTime() { return this.burstTime; }
-	public float getExecuted() { return this.executed; }
-	public float getTurnaroundTime() { return this.turnaroundTime; }
+	public int getArrivalTime() { return this.arrivalTime; }
+	public int getBeginTime() { return this.beginTime; }
+	public int getBurstTime() { return this.burstTime; }
+	public int getExecuted() { return this.executed; }
+	public int getTurnaroundTime() { return this.turnaroundTime; }
 	public String getName() { return this.name; }
-	public float getRemainingTime() { return this.remainingTime; }
-	public float getWaitTime() { return this.waitTime; }
-	public float getResponseTime() { return this.responseTime; }
-	public float getCompletionTime() { return this.completionTime; }
-	public void setRemainingTime(float remainingTime) { this.remainingTime = remainingTime; }
-	public void setWaitTime(float waitTime) { this.waitTime = waitTime; }
-	public void setBurstTime(float burstTime) { this.burstTime = burstTime; }
-	public void setArrivalTime(float arrivalTime) { this.arrivalTime = arrivalTime; }
-	public void setBeginTime(float beginTime) { this.beginTime = beginTime; }
-	public void setExecuted(float executed) { this.executed = executed; }
-	public void setTurnaroundTime(float turnaroundTime) { this.turnaroundTime = turnaroundTime; }
-	public void setResponseTime(float responseTime) { this.responseTime = responseTime; }
-	public void setCompletionTime(float completionTime) { this.completionTime = completionTime; }
+	public int getRemainingTime() { return this.remainingTime; }
+	public int getWaitTime() { return this.waitTime; }
+	public int getResponseTime() { return this.responseTime; }
+	public int getCompletionTime() { return this.completionTime; }
+	public void setRemainingTime(int remainingTime) { this.remainingTime = remainingTime; }
+	public void setWaitTime(int waitTime) { this.waitTime = waitTime; }
+	public void setBurstTime(int burstTime) { this.burstTime = burstTime; }
+	public void setArrivalTime(int arrivalTime) { this.arrivalTime = arrivalTime; }
+	public void setBeginTime(int beginTime) { this.beginTime = beginTime; }
+	public void setExecuted(int executed) { this.executed = executed; }
+	public void setTurnaroundTime(int turnaroundTime) { this.turnaroundTime = turnaroundTime; }
+	public void setResponseTime(int responseTime) { this.responseTime = responseTime; }
+	public void setCompletionTime(int completionTime) { this.completionTime = completionTime; }
 	
 	
-	public PCB(String name, float burstTime, float arrivalTime) {
+	public PCB(String name, int burstTime, int arrivalTime) {
 		this.name = name;
 		this.burstTime = burstTime;
 		this.remainingTime = burstTime;
